+++ documentation/docs/user/changelog.md
@@ -7,6 +7,7 @@
 - Referencing jobs in the connector editor are now ordered alphabetically.
 - Errors during simulated job executions are now shown in the job navigator.
 - Simulation results of the last action are added as 'stale' to newly created actions.
+- Simulation results are now shown in an accordion component.
 
 ## Version 0.2.5
 - Added 'mustache-selector' for JSON simulation results.